fix(cosh-ng): [shell] wrap styled panels safely
Review follow-ups on the panel rendering commit:

- restore wrapping in rich_styled_block_lines with Wrap { trim: false }:
  keeps leading indentation intact for pre-wrapped callers (help panel)
  while unwrapped styled markdown paragraphs and list items wrap again
  instead of being truncated on narrow styled terminals (P1)
- respect styles_enabled() in the styled channel so NO_COLOR and
  non-TTY output carries no ANSI escapes (P2)
- pin TERM=xterm-256color in the raw_cli /help reference test so
  rich-border assertions do not depend on the host terminal (P2)
- tighten HelpPanelModel/Group/Entry and write_help_panel to pub(crate)
  per the crate-internal visibility guideline (P3)
- move the new help/notice panel tests to ui/agent_render/help_tests.rs
  and the /recommendations tests to slash/recommendations_tests.rs so
  oversized inline test modules stop growing (P2/P3)
- regression tests: styled long-paragraph tail retention, ANSI-free
  output with styles disabled
